:recycle: refactor: 멀티 모듈을 고려한 디레토리 구조 정리
