+++ backend/src/main/java/kernel360/techpick/core/config/JpaAuditingConfig.java
@@ -1,0 +1,9 @@
+package kernel360.techpick.core.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfig {
+}
