+++ backend/src/main/java/kernel360/techpick/feature/user/domain/readme.md
@@ -1,0 +1,2 @@
+디렉토리 유지를 위한 Dummy Readme 입니다.
+(파일 없으면 디렉토리 git에 추가 안되서 함)
